be/pyproject.toml
[tool.poetry]
name = "be"
version = "0.1.0"
description = ""
authors = ["Erikyy <erikyl@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
nanoid = "^2.0.0"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
django-cors-headers = "^3.12.0"
Pillow = "^9.1.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
